Allow arcade drive implementations provide custom quick turn triggers
Since quick turn triggers may be based off of throttle and/or turn
values, everything is now provided using a single method call.

Any decisions on how the quick turn should behave based on the current
input values is still in the hands of the base class. This detail is
bound to the physical attributes of the Robot and should not vary based
on inputs schemes.