Add in test mode to determine polarity of individual motors
Since new motor controllers were installed, motor polarity needs to be
tested. To simplify this, a new test mode was added which runs one motor
at a time at 50% forward power.

To get reliable timings in test mode, the robot has been switched to be
a `TimedRobot` instead of an `IterativeRobot`.