Fix polarity of CAN-controlled motors
With the test mode, the polarity of all motors was tested and set
correctly. This exposed the fact that *all* motors were inverted in the
previous iteration (pre-CAN) of the code. Fixing this caused the driving
controls to run backwards so signs were inverted to compensate in both
of the "Tiger Drive" drive control schemes. Documentation within the
`TigerDriveInput` base class was beefed up to fill in the gaps exposed
by this exercise.